fix(avatar): cast the aspect-ratio scale to float for strict operand mode
Psalm runs with strictBinaryOperands, where multiplying an int width by a
value it infers as int|float is an error rather than an implicit promotion.
The scale is a ratio and is only ever meaningfully fractional, so state that
in the expression instead of relying on inference from the division. No
behaviour change: the surrounding round() and int casts already produced a
float multiplication at runtime.

+++ app/lib/Service/AvatarSyncService.php
@@ -339,7 +339,7 @@ class AvatarSyncService {
 			$side = min(max($width, $height), self::MAX_DIMENSION);
 
 			// Preserve the aspect ratio inside that square.
-			$scale = min($side / $width, $side / $height);
+			$scale = (float)min($side / $width, $side / $height);
 			$drawWidth = max(1, (int)round((float)$width * $scale));
 			$drawHeight = max(1, (int)round((float)$height * $scale));
